# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "OUTLIER API",
  "version": "0.1.0"
 },
 "paths": {
  "/api/listings": {
   "post": {
    "tags": [
     "listings"
    ],
    "summary": "Create Listing",
    "operationId": "create_listing_api_listings_post",
    "parameters": [
     {
      "name": "fetch_images",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Fetch Images"
      }
     },
     {
      "name": "analyze",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Analyze"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ListingIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "listings"
    ],
    "summary": "List Listings",
    "operationId": "list_listings_api_listings_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "domain",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Domain"
      }
     },
     {
      "name": "tier",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Tier"
      }
     },
     {
      "name": "source",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Source"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "include_demo",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Include Demo"
      }
     },
     {
      "name": "include_archived",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Include Archived"
      }
     },
     {
      "name": "watchlist_only",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Watchlist Only"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "score",
       "title": "Sort"
      }
     },
     {
      "name": "order",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "desc",
       "title": "Order"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/import/json": {
   "post": {
    "tags": [
     "listings"
    ],
    "summary": "Import Json",
    "operationId": "import_json_api_listings_import_json_post",
    "parameters": [
     {
      "name": "fetch_images",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Fetch Images"
      }
     },
     {
      "name": "analyze",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Analyze"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true,
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/import/csv": {
   "post": {
    "tags": [
     "listings"
    ],
    "summary": "Import Csv",
    "operationId": "import_csv_api_listings_import_csv_post",
    "parameters": [
     {
      "name": "fetch_images",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Fetch Images"
      }
     },
     {
      "name": "analyze",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Analyze"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_csv_api_listings_import_csv_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/import/email": {
   "post": {
    "tags": [
     "listings"
    ],
    "summary": "Import Email",
    "description": "Import ShopGoodwill Personal Shopper notification e-mails (.eml upload or pasted HTML).",
    "operationId": "import_email_api_listings_import_email_post",
    "parameters": [
     {
      "name": "fetch_images",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Fetch Images"
      }
     },
     {
      "name": "analyze",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Analyze"
      }
     }
    ],
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_email_api_listings_import_email_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/import/page": {
   "post": {
    "tags": [
     "listings"
    ],
    "summary": "Import Page",
    "description": "Import a saved ShopGoodwill item page (HTML).",
    "operationId": "import_page_api_listings_import_page_post",
    "parameters": [
     {
      "name": "fetch_images",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Fetch Images"
      }
     },
     {
      "name": "analyze",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Analyze"
      }
     }
    ],
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_page_api_listings_import_page_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}": {
   "get": {
    "tags": [
     "listings"
    ],
    "summary": "Get Listing",
    "operationId": "get_listing_api_listings__listing_id__get",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "listings"
    ],
    "summary": "Patch Listing",
    "operationId": "patch_listing_api_listings__listing_id__patch",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ListingPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "listings"
    ],
    "summary": "Delete Listing",
    "operationId": "delete_listing_api_listings__listing_id__delete",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/snapshot": {
   "post": {
    "tags": [
     "listings"
    ],
    "summary": "Add Snapshot",
    "description": "Record a manual price/status update (authorized source: you looked at the page).",
    "operationId": "add_snapshot_api_listings__listing_id__snapshot_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SnapshotIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/images": {
   "post": {
    "tags": [
     "listings"
    ],
    "summary": "Upload Images",
    "operationId": "upload_images_api_listings__listing_id__images_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_images_api_listings__listing_id__images_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/image-urls": {
   "post": {
    "tags": [
     "listings"
    ],
    "summary": "Add Image Urls",
    "operationId": "add_image_urls_api_listings__listing_id__image_urls_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ImageUrlsIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/images/{image_id}": {
   "delete": {
    "tags": [
     "listings"
    ],
    "summary": "Delete Image",
    "operationId": "delete_image_api_listings__listing_id__images__image_id__delete",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     },
     {
      "name": "image_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Image Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/images/{image_id}": {
   "get": {
    "tags": [
     "images"
    ],
    "summary": "Serve Image",
    "operationId": "serve_image_api_images__image_id__get",
    "parameters": [
     {
      "name": "image_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Image Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/listings/{listing_id}/finance": {
   "post": {
    "tags": [
     "comparables"
    ],
    "summary": "Finance",
    "operationId": "finance_api_listings__listing_id__finance_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FinanceQuery"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/webhooks/listing": {
   "post": {
    "tags": [
     "misc"
    ],
    "summary": "Webhook Listing",
    "operationId": "webhook_listing_api_webhooks_listing_post",
    "parameters": [
     {
      "name": "analyze",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Analyze"
      }
     },
     {
      "name": "x-outlier-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Outlier-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ListingIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_import_csv_api_listings_import_csv_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_import_csv_api_listings_import_csv_post"
   },
   "Body_import_email_api_listings_import_email_post": {
    "properties": {
     "file": {
      "anyOf": [
       {
        "type": "string",
        "contentMediaType": "application/octet-stream"
       },
       {
        "type": "null"
       }
      ],
      "title": "File"
     },
     "html": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Html"
     }
    },
    "type": "object",
    "title": "Body_import_email_api_listings_import_email_post"
   },
   "Body_import_page_api_listings_import_page_post": {
    "properties": {
     "file": {
      "anyOf": [
       {
        "type": "string",
        "contentMediaType": "application/octet-stream"
       },
       {
        "type": "null"
       }
      ],
      "title": "File"
     },
     "html": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Html"
     },
     "url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Url"
     }
    },
    "type": "object",
    "title": "Body_import_page_api_listings_import_page_post"
   },
   "Body_upload_images_api_listings__listing_id__images_post": {
    "properties": {
     "files": {
      "items": {
       "type": "string",
       "contentMediaType": "application/octet-stream"
      },
      "type": "array",
      "title": "Files"
     }
    },
    "type": "object",
    "required": [
     "files"
    ],
    "title": "Body_upload_images_api_listings__listing_id__images_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ImageUrlsIn": {
    "properties": {
     "urls": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Urls"
     }
    },
    "type": "object",
    "required": [
     "urls"
    ],
    "title": "ImageUrlsIn"
   },
...
   "ListingIn": {
    "properties": {
     "source": {
      "type": "string",
      "title": "Source",
      "default": "manual"
     },
     "source_item_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Item Id"
     },
     "source_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Url"
     },
     "extraction_method": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Extraction Method"
     },
     "title": {
      "type": "string",
      "maxLength": 500,
      "minLength": 1,
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "seller": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seller"
     },
     "domain": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Domain"
     },
     "current_bid": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Current Bid"
     },
     "num_bids": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Num Bids"
     },
     "buy_now_price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Buy Now Price"
     },
     "ends_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ends At"
     },
     "shipping_cost": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shipping Cost"
     },
     "handling_fee": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Handling Fee"
     },
     "other_costs": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Other Costs"
     },
     "measurements": {
      "additionalProperties": true,
      "type": "object",
      "title": "Measurements"
     },
     "condition_text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Condition Text"
     },
     "status": {
      "type": "string",
      "title": "Status",
      "default": "active"
     },
     "image_urls": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Image Urls"
     },
     "raw": {
      "additionalProperties": true,
      "type": "object",
      "title": "Raw"
     },
     "is_demo": {
      "type": "boolean",
      "title": "Is Demo",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "title"
    ],
    "title": "ListingIn",
    "description": "Normalized listing payload used by every ingestion path."
   },
...
  }
 }
}